Update and rename ConcurrentHashMapTest.java to Issue1142.java
