--- checker/tests/nullness/ConcurrentHashMapTest.java
@@ -1,16 +1,0 @@
-
-import java.util.concurrent.ConcurrentHashMap;
-import org.checkerframework.checker.nullness.qual.Nullable;
-import org.checkerframework.checker.nullness.qual.NonNull;
-
-class Issue1142 {
-
-    void foo() {
-
-        ConcurrentHashMap<Integer, @Nullable Integer> chm = new ConcurrentHashMap<Integer, @Nullable Integer>();
-        chm.put(1, 1);
-        chm.put(3, 3);
-        chm.put(4, null);
-
-    }
-}
